Updated SimpleCacheTest.php

+++ src/SimpleCache.php
@@ -3,34 +3,16 @@ declare(strict_types=1);
 
 namespace ItalyStrap\Cache;
 
-use ArrayObject;
 use DateInterval;
-use DateTime;
 use Exception;
-use Fig\Cache\KeyValidatorTrait;
 use ItalyStrap\Cache\Exceptions\SimpleCacheInvalidArgumentException;
 use ItalyStrap\Storage\StorageInterface;
-use ItalyStrap\Storage\Transient;
 use ItalyStrap\Tests\ConvertDateIntervalToIntegerTrait;
 use Psr\SimpleCache\CacheInterface as PsrSimpleCacheInterface;
-use Traversable;
-use function array_keys;
-use function get_class;
-use function gettype;
-use function intval;
-use function is_array;
-use function is_object;
-use function is_string;
-use function iterator_to_array;
-use function sprintf;
 
-/**
- * @psalm-api
- */
 class SimpleCache implements PsrSimpleCacheInterface {
 
-//	use KeyValidatorTrait;
-	use ConvertDateIntervalToIntegerTrait, ToArrayTrait;
+	use ConvertDateIntervalToIntegerTrait, ToArrayTrait, KeyValidatorTrait;
 
 	private StorageInterface $storage;
 	private array $used_keys = [];
@@ -39,36 +21,18 @@ class SimpleCache implements PsrSimpleCacheInterface {
 		$this->storage = $storage;
 	}
 
-	/**
-	 * @inheritDoc
-	 */
 	public function has( $key ): bool {
 		return $this->get( $key, false ) !== false;
 	}
 
-	/**
-	 * @inheritDoc
-	 * If you need to store booleans use 0 or 1 because
-	 * get_transient() return false if value is not set or is expired
-	 * @see \get_transient()
-	 */
 	public function get( $key, $default = null ) {
 		$this->assertKeyIsValid( $key );
 		$this->addUsedKey( $key );
 
-		/** @var mixed $value */
 		$value = $this->storage->get( $key );
-		if ( false === $value ) {
-			$value = $default;
-		}
-
-		return $value;
+		return $value ?: $default;
 	}
 
-	/**
-	 * @inheritDoc
-	 * @throws Exception
-	 */
 	public function set( $key, $value, $ttl = null ): bool {
 		$this->assertKeyIsValid( $key );
 		$this->addUsedKey( $key );
@@ -77,49 +41,42 @@ class SimpleCache implements PsrSimpleCacheInterface {
 			$ttl = $this->convertDateIntervalToInteger($ttl);
 		}
 
-		return $this->storage->set( $key, $value, (int)$ttl );
+		return $this->storage->set( (string)$key, \is_object($value) ? clone $value : $value, (int)$ttl );
 	}
 
-	/**
-	 * @inheritDoc
-	 */
 	public function delete( $key ): bool {
 		$this->assertKeyIsValid( $key );
+
+		if (!\array_key_exists($key, $this->usedKeys())) {
+			return true;
+		}
+
 		$this->deleteUsedKey( $key );
 		return $this->storage->delete( $key );
 	}
 
-	/**
-	 * @inheritDoc
-	 */
-	public function getMultiple( $keys, $default = null ) {
+	public function getMultiple( $keys, $default = null ): iterable {
 		if (!\is_iterable($keys)) {
 			throw new SimpleCacheInvalidArgumentException( 'Cache keys must be array or Traversable' );
 		}
-		$keys = $this->assertKeysAreValid( $keys );
 
-		$data = [];
+		$gen = function () use ($keys, $default) {
+			foreach ($keys as $key) {
+				yield $key => $this->get( $key, $default );
+			}
+		};
 
-		foreach ( $keys as $key ) {
-			$data[ $key ] = $this->get( $key, $default );
-		}
-
-		return $data;
+		return $gen();
 	}
 
-	/**
-	 * @inheritDoc
-	 */
 	public function setMultiple( $values, $ttl = null ): bool {
 		if (!\is_iterable($values)) {
 			throw new SimpleCacheInvalidArgumentException( 'Cache values must be array or Traversable' );
 		}
 
-		$values = $this->toArray($values, 'values');
-
 		$success = true;
 		foreach ( $values as $key => $value ) {
-			if ( $this->set( $key, $value, $ttl ) ) {
+			if ( $this->set($key, $value, $ttl ) ) {
 				continue;
 			}
 			$success = false;
@@ -128,14 +85,11 @@ class SimpleCache implements PsrSimpleCacheInterface {
 		return $success;
 	}
 
-	/**
-	 * @inheritDoc
-	 */
 	public function deleteMultiple( $keys ): bool {
 		if (!\is_iterable($keys)) {
 			throw new SimpleCacheInvalidArgumentException( 'Cache keys must be array or Traversable' );
 		}
-		$keys = $this->assertKeysAreValid( $keys );
+
 		$success = true;
 		/** @var string $key */
 		foreach ( $keys as $key ) {
@@ -148,58 +102,26 @@ class SimpleCache implements PsrSimpleCacheInterface {
 		return $success;
 	}
 
-	/**
-	 * @psalm-suppress InvalidThrow
-	 * @inheritDoc
-	 * @throws \Psr\SimpleCache\InvalidArgumentException
-	 */
 	public function clear(): bool {
 		return $this->deleteMultiple( $this->usedKeys() );
 	}
 
-	/**
-	 * @psalm-suppress ReservedWord
-	 * @param mixed $key
-	 */
 	private function assertKeyIsValid( $key ): void {
-		if ( ! is_string( $key ) ) {
-			throw new SimpleCacheInvalidArgumentException( sprintf(
-				'The $key must be a string, %s given',
-				gettype( $key )
-			) );
-		}
-
-		if ( empty( $key ) ) {
-			throw new SimpleCacheInvalidArgumentException( 'The $key must be not empty' );
+		try {
+			$this->validateKey($key);
+		} catch (\InvalidArgumentException $e) {
+			throw new SimpleCacheInvalidArgumentException($e->getMessage(), $e->getCode());
 		}
 	}
 
-	/**
-	 * @param array|iterable $keys
-	 * @return array<string>
-	 */
-	private function assertKeysAreValid( $keys ): array {
-		$keys = $this->toArray($keys, 'keys');
-		return $keys;
-	}
-
-	/**
-	 * @param string $key
-	 */
 	private function addUsedKey( $key ): void {
 		$this->used_keys[ $key ] = $key;
 	}
 
-	/**
-	 * @param string $key
-	 */
 	private function deleteUsedKey( $key ): void {
 		unset( $this->used_keys[ $key ] );
 	}
 
-	/**
-	 * @return array
-	 */
 	private function usedKeys(): array {
 		return $this->used_keys;
 	}
